add shadcn demo (#517)
# Why

Shadcn CLI no longer aids in the setup of Expo Router projects since it
isn't one of the recognized frameworks. So I'm adding a slightly
opinionated example to make all the requirements (things like postcss,
tailwind, autoprefixer, etc) easy to setup.

This demo is web-only and won't run on native. We can add native shims
in SDK 52 with DOM components.